Implement searching of lshw
In order to use lshw library a (tmp) file for lshw.xml must be created.

As the file destination for ZipFile.extract must be empty, a tempdir
is created and then a hard path is set for lshw.xml within there.

The use of the lshw library is unsatisfactory as much functionality
is missing & must be monkeypatched in, however it will have to do for
now due to time constraints. Expect this methodology to be overhauled.

Additionally all code has been moved into lib