Add output capability
The output location is set as a fixed file within /opt/
If the location of a template is passed with the `-t` cli argument
then  the output takes the form of that template parsed using erubis.
Else the output is parsed as yaml and appended to the existing file if
it exists
The hash has been re-organised to not use the node name as a single
top-level key to make the eRuby prettier. If output is yaml this is
reinstated to prettify the result