Store filled templates uniquely
Each is given a name of `NODE_NAME+TEMPLATE_NAME` within the output
directory.
This is so they're kept permanantly, so they don't overwrite the
yaml, so they're seperated and so a new version of the same template
will override itself.
Also TARGET_FILE has been changed to YAML_FILE (as this is now
accurately its purpose) and a new OUTPUT_DIR constant has been defined