Add os mapping capability
To prevent needing to redefine templates based on the target node's OS
  (as different node OS presents different lshw output)
Create hash to store various feilds based on OS
OS is currently passed as a CLI option that is neccesary if a template
is being used. This will likely change to being scrubbed when more input
data is added.