Move node related code to utils
Extract code for dealing with --all and --group options and move it to
new `resolve_node_options` method in utils
Also create `select_nodes` method for calling `find_nodes`,
`find_all_nodes` and `find_nodes_in_groups` methods and move all to
utils
This additionally required the moving of the utils methods to the
Inventoryware module as they reference Inventoryware constants.
The utils module should have already been in this module. Additionally
it was placed in a `Utils` module with `Inventoryware`
All the util methods were converted to class methods with `self.`