Merge with existing data when parsing
If there's already data on a node when another .zip is parsed for it the
two sets of data will be merged - with the new data getting priority
for overrides - and then the combination will be saved

It is possible that this could cause problems when replacing lshw
output. The nested hash could become invalid when recursively combined
with another. An exception may be made for this - more research
incoming.