Determine if a node's been imported when rendering
Rather than storing that information explicitly iin the hash, as was the
old system.
This check doesn't stricly speaking check for if an import has been
performed but instead checks for the headings that can be expected for
all nodes that have been imported. This should be sufficient for all use
cases.