Create and use 'node' model
As the util methods for manipulating yaml (read_node_yaml,
read_node_or_create, output_node_yaml) were all tied to nodes as an
entity and as such it is clearer and more OOP-y to have them belong to a
class describing a node. All these methods have thus been moved to the
new node.rb (as open, create_if_non_existent, and save respectively)

A node has 'data', 'name' and 'location' as accesible attributes.
If @data hasn't yet got a value open will be called to provide it with
the node's yaml. This will cause issues if called on nodes that don't
exist.