Allow specifying single names that are ambiguous
Fixes bug whereby it was impossible to specify a file name if their
existed longer versions of that name as it would just ask for
confirmation.
e.g. if node & node1 existed in the system and 'node' was specified, a
confirmation message would be displayed asking you to refine your search
with no wy to actually select 'node'