Add initial prompt for adding map metadata
This is a test of tty-prompt that will be used to ask the user if they
wish to add map metadata or not when running `modify map` on an asset.
For now this will use no as the default so the user can simply press
enter to skip adding metadata for an asset's map.