Add checks for schemas
Was somewhat torn on where to put these - in the end I went for in
multi_node_command and single_node_command as well as in find_all_nodes.

The latter is to ensure a warning is produced when using 'list'.

It will need to be ensured going forward that no commands circumvent
these measures