Refactor to call new migration scripts
Now that migrations are stored within `scripts/migrations` I refactored
this method to act as the new central method for schema migration. It
will check the given asset's schema against the schema found within
`version.rb`.

If necessary it will perform any and all migrations necessary, in order,
to get it to the correct schema version.

Most of the errors etc have remained the same but moved slightly.