Fix CPU helper not correctly identifying CPUs...
The CPU handler would previously look for "class: processor" in the
node dmidecode output.

When a SATA controller is present it will have hardware seen by the node
as having a "class: processor" meaning that this is included in the CPUs
hash.

This modification to the handler uses the "description" field instead
which will ensure that it only takes items described as CPUs.