Remove the check around the yaml_dir
There is no point having the directory handling code for a single
`yaml_dir`. It should just be created by initialize

Technically it could be a file which will trigger a generic file
path error. In practice this will never happen due to how the ruby
app is packaged. And if it did, it's easily fixed