Replace rvim with the default $EDITOR behaviour
This was an old fix to an injection attack that isn't possible
within this code base alone

(the app use to be for a REPL running with root level privileges)

However this creates all sorts of issues if `rvim` isn't installed.
Instead $EDITOR should be used (aka the TTY::Editor) default.

Any higher order concerns about injection attacks should be handled
in that application explicitly. AKA set $EDITOR to rvim